+++ optimizator/function/domain_sampler.py
@@ -4,4 +4,4 @@ import jax.random as random
 def domain_sampler(nb_points, dims, key):
     magnitude = 1.0
     key = random.PRNGKey(key)
-    return magnitude * 2.0 * (random.uniform(key, (nb_points, dims)) - 0.5)
+    return magnitude * 2.0 * (random.uniform(key, (nb_points, dims), dtype=jnp.float64) - 0.5)
